requirements.txt
# Core simulation & RL
gymnasium>=0.29.0
pybullet>=3.2.6
stable-baselines3>=2.3.0
shimmy>=1.3.0

# Deep learning backend
torch>=2.2.0
torchvision>=0.17.0

# Utilities
numpy>=1.26.0
pyyaml>=6.0.1

# Logging & visualisation (optional but recommended)
tensorboard>=2.16.0
matplotlib>=3.8.0

# Testing
pytest>=8.0.0
